+++ generators/src/helpers.js
@@ -1,5 +1,0 @@
-function betterStringSort(a, b) {
-  return a.toLowerCase().localeCompare(b.toLowerCase())
-}
-
-module.exports = { betterStringSort }