+++ generators/src/plop-interfaces.ts
@@ -2,6 +2,7 @@ export interface Answers {
   componentName: string;
   propName: string;
   ruleName: string;
+  referenceRepo: string;
   referencePR: string;
   message?: string;
 }