+++ packages/eslint-plugin-pf-codemods/src/rules/v6/mastheadNameChanges/mastheadNameChangesInput.tsx
@@ -1,0 +1,3 @@
+import { MastheadBrand } from "@patternfly/react-core";
+
+export const MastheadNameChanges = () => <MastheadBrand>Logo</MastheadBrand>;