+++ packages/eslint-plugin-pf-codemods/src/rules/v6/mastheadNameChanges/mastheadNameChangesOutput.tsx
@@ -1,0 +1,3 @@
+import { MastheadLogo } from "@patternfly/react-core";
+
+export const MastheadNameChanges = () => <MastheadLogo data-codemods>Logo</MastheadLogo>;